requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.3.0
djangorestframework==3.15.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
requests==2.32.0
sqlparse==0.5.5
urllib3==2.6.3

Pillow
djangorestframework-simplejwt==5.3.1
django-allauth
google-auth
drf-yasg
python-jose
